+++ data-raw/generate_rebase_database.R
@@ -15,5 +15,8 @@ if (exists("ggchord_common_feature_database", envir = old, inherits = FALSE)) {
     "ggchord_common_feature_database", envir = old
   )
 }
+if (exists("ggchord_primer_database", envir = old, inherits = FALSE)) {
+  objects$ggchord_primer_database <- get("ggchord_primer_database", envir = old)
+}
 list2env(objects, envir = environment())
 save(list = names(objects), file = "R/sysdata.rda", compress = "xz")